fruit-app-streamlit/fruit-app/requirements.txt

streamlit
ultralytics
opencv-python-headless
Pillow
streamlit-authenticator
PyYAML
